# [target-allocator] Update scrape config upon discovery config apply