--- a/cmd/otel-allocator/target/discovery_test.go
+++ b/cmd/otel-allocator/target/discovery_test.go
@@ -294,6 +294,7 @@ type mockScrapeConfigUpdater struct {
 	mockCfg map[string]*promconfig.ScrapeConfig
 }
 
-func (m *mockScrapeConfigUpdater) UpdateScrapeConfigResponse(cfg map[string]*promconfig.ScrapeConfig) {
+func (m *mockScrapeConfigUpdater) UpdateScrapeConfigResponse(cfg map[string]*promconfig.ScrapeConfig) error {
 	m.mockCfg = cfg
+	return nil
 }