Fixes #1359.

This is a proposed alternative to solve the data race issue in scrape config handler and at the same time now blowing up the memory on a repeated handler call.

The stored scrape config response gets updated only upon relevant changes in the `Discoverer`. When a change is detected, it calls a method to update the stored scrape config in the HTTP server response. All responses to scrape config handler are then served from this stored JSON response (or alternatively an error is returned if the latest config update failed). We also keep the hash of config within the discoverer, to prevent sending unnecessarily scrape configs over channel if no update has been detected.

Benchmarks from my machine:
```
goos: darwin
goarch: arm64
pkg: github.com/open-telemetry/opentelemetry-operator/cmd/otel-allocator/server
                                    │     old.txt     │            new_sync.txt             │
                                    │     sec/op      │    sec/op     vs base               │
ScrapeConfigsHandler/0_targets-10        1.929µ ± ∞ ¹   1.880µ ± ∞ ¹   -2.54% (p=0.032 n=5)
ScrapeConfigsHandler/5_targets-10      240.840µ ± ∞ ¹   2.107µ ± ∞ ¹  -99.13% (p=0.008 n=5)
ScrapeConfigsHandler/10_targets-10     452.602µ ± ∞ ¹   2.543µ ± ∞ ¹  -99.44% (p=0.008 n=5)
ScrapeConfigsHandler/50_targets-10    2246.603µ ± ∞ ¹   3.737µ ± ∞ ¹  -99.83% (p=0.008 n=5)
ScrapeConfigsHandler/100_targets-10   4672.521µ ± ∞ ¹   6.042µ ± ∞ ¹  -99.87% (p=0.008 n=5)
ScrapeConfigsHandler/500_targets-10   23730.46µ ± ∞ ¹   18.43µ ± ∞ ¹  -99.92% (p=0.008 n=5)
geomean                                  611.7µ         4.015µ        -99.34%
¹ need >= 6 samples for confidence interval at level 0.95

                                    │     old.txt     │             new_sync.txt             │
                                    │      B/op       │     B/op       vs base               │
ScrapeConfigsHandler/0_targets-10       6.376Ki ± ∞ ¹   6.423Ki ± ∞ ¹   +0.74% (p=0.008 n=5)
ScrapeConfigsHandler/5_targets-10     206.839Ki ± ∞ ¹   8.110Ki ± ∞ ¹  -96.08% (p=0.008 n=5)
ScrapeConfigsHandler/10_targets-10     311.29Ki ± ∞ ¹   10.36Ki ± ∞ ¹  -96.67% (p=0.008 n=5)
ScrapeConfigsHandler/50_targets-10    1651.12Ki ± ∞ ¹   24.36Ki ± ∞ ¹  -98.52% (p=0.008 n=5)
ScrapeConfigsHandler/100_targets-10   4117.17Ki ± ∞ ¹   46.37Ki ± ∞ ¹  -98.87% (p=0.008 n=5)
ScrapeConfigsHandler/500_targets-10   22878.4Ki ± ∞ ¹   182.4Ki ± ∞ ¹  -99.20% (p=0.008 n=5)
geomean                                 632.2Ki         21.93Ki        -96.53%
¹ need >= 6 samples for confidence interval at level 0.95

                                    │     old.txt     │            new_sync.txt            │
                                    │    allocs/op    │  allocs/op   vs base               │
ScrapeConfigsHandler/0_targets-10         24.00 ± ∞ ¹   23.00 ± ∞ ¹   -4.17% (p=0.008 n=5)
ScrapeConfigsHandler/5_targets-10       3396.00 ± ∞ ¹   23.00 ± ∞ ¹  -99.32% (p=0.008 n=5)
ScrapeConfigsHandler/10_targets-10      6702.00 ± ∞ ¹   23.00 ± ∞ ¹  -99.66% (p=0.008 n=5)
ScrapeConfigsHandler/50_targets-10     33144.00 ± ∞ ¹   23.00 ± ∞ ¹  -99.93% (p=0.008 n=5)
ScrapeConfigsHandler/100_targets-10    66208.00 ± ∞ ¹   23.00 ± ∞ ¹  -99.97% (p=0.008 n=5)
ScrapeConfigsHandler/500_targets-10   330516.00 ± ∞ ¹   23.00 ± ∞ ¹  -99.99% (p=0.008 n=5)
geomean                                  8.570k         23.00        -99.73%
¹ need >= 6 samples for confidence interval at level 0.95
```