@matej-g 2023-02-01
I wonder now if there is actually need to do this over channel rather than doing this synchronously and just update the scrape config response directly from `Discoverer`