@jaronoff97 2023-02-01
yeah i'm not sure we need the channel here, you can do as we've done with [interfaces like the hook](https://github.com/open-telemetry/opentelemetry-operator/blob/main/cmd/otel-allocator/target/discovery.go#L45-L47) and just pass that around. Could you also add some tests for the UpdateScrapeConfigResponse? They'll probably look similar to the ones removed in the `TestScrapeConfigsHandler_Hashing`