@matej-g 2023-02-01
Yeah, I guess no reason to complicate it here, let me do that + add tests and come back with more commits